Update docker commands and add patch instructions in README
This update includes a numbered version for the docker image build and new port mappings in the docker run command. It also adds instructions on how to apply patches to WebLogic server in the README file. Changes have also been made to the user oracle setup and entry point script in the setup.sh file.